# Fix reassembly pages allocation/shrinking (second attempt)