--- a/reassembly/memory.go
+++ b/reassembly/memory.go
@@ -23,58 +23,17 @@ var memLog = flag.Bool("assembly_memuse_log", defaultDebug, "If true, the github
 // pageCache is a concurrency-unsafe store of page objects we use to avoid
 // memory allocation as much as we can.
 type pageCache struct {
-	free         []*page
-	pcSize       int
-	size, used   int
+	pagePool     *sync.Pool
+	used         int
 	pageRequests int64
-	ops          int
-	nextShrink   int
 }
-
-const initialAllocSize = 1024
 
 func newPageCache() *pageCache {
 	pc := &pageCache{
-		free:   make([]*page, 0, initialAllocSize),
-		pcSize: initialAllocSize,
-	}
-	pc.grow()
+		pagePool: &sync.Pool{
+			New: func() interface{} { return new(page) },
+		}}
 	return pc
-}
-
-// grow exponentially increases the size of our page cache as much as necessary.
-func (c *pageCache) grow() {
-	pages := make([]page, c.pcSize)
-	c.size += c.pcSize
-	for i := range pages {
-		c.free = append(c.free, &pages[i])
-	}
-	if *memLog {
-		log.Println("PageCache: created", c.pcSize, "new pages, size:", c.size, "cap:", cap(c.free), "len:", len(c.free))
-	}
-	// control next shrink attempt
-	c.nextShrink = c.pcSize
-	c.ops = 0
-	// prepare for next alloc
-	c.pcSize *= 2
-}
-
-// Remove references to unused pages to let GC collect them
-// Note: memory used by c.free itself it not collected.
-func (c *pageCache) tryShrink() {
-	var min = c.pcSize / 2
-	if min < initialAllocSize {
-		min = initialAllocSize
-	}
-	if len(c.free) <= min {
-		return
-	}
-	for i := range c.free[min:] {
-		c.free[min+i] = nil
-	}
-	c.size -= len(c.free) - min
-	c.free = c.free[:min]
-	c.pcSize = min
 }
 
 // next returns a clean, ready-to-use page object.
@@ -82,24 +41,15 @@ func (c *pageCache) next(ts time.Time) (p *page) {
 	if *memLog {
 		c.pageRequests++
 		if c.pageRequests&0xFFFF == 0 {
-			log.Println("PageCache:", c.pageRequests, "requested,", c.used, "used,", len(c.free), "free")
+			log.Println("PageCache:", c.pageRequests, "requested,", c.used, "used,")
 		}
 	}
-	if len(c.free) == 0 {
-		c.grow()
-	}
-	i := len(c.free) - 1
-	p, c.free = c.free[i], c.free[:i]
+	p = c.pagePool.Get().(*page)
 	p.seen = ts
 	p.bytes = p.buf[:0]
 	c.used++
 	if *memLog {
 		log.Printf("allocator returns %s\n", p)
-	}
-	c.ops++
-	if c.ops > c.nextShrink {
-		c.ops = 0
-		c.tryShrink()
 	}
 
 	return p
@@ -113,7 +63,7 @@ func (c *pageCache) replace(p *page) {
 	}
 	p.prev = nil
 	p.next = nil
-	c.free = append(c.free, p)
+	c.pagePool.Put(p)
 }
 
 /*
@@ -145,6 +95,8 @@ type StreamPool struct {
 	nextAlloc          int
 	newConnectionCount int64
 }
+
+const initialAllocSize = 1024
 
 func (p *StreamPool) grow() {
 	conns := make([]connection, p.nextAlloc)