@gconnell 2021-03-04
I think this issue is surprising folks enough that a small hit in performance is unfortunately warranted, and I really appreciate the work you've done here for the fix.  Going to merge this as is, and thanks so much for the time and effort put into it.

I have one potential idea which may increase performance slightly:  it would be interesting to see if pre-allocating some # in New() impacts performance much, for example:

```
  pool := &sync.Pool{}
  pool.New = func() interface{} {
    pages := make([]page, 8)  // Make some number of them at once, with a single alloc
    for _, p := range pages[1:] {
      pool.Put(&p)  // Make most of them available to subsequent Get calls
    }
    return &p[0]  // Return one that we made
  }
```

I've done no testing on if this is possible, and it still will utilize a _little_ more memory, because it has to wait until the block of 8 pages are all freed before releasing the memory.  But it may still provide the majority of benefit for this issue while speeding things up a bit.